requirements.txt
click==8.0.1
cmake==3.20.5
colorama==0.4.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
install==1.3.4
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
